This initial solution was based on solving the labyrinth using the left hand rule.  This was used because only a small portion of the labyrinth is originally known.  To mark the way back, we kept track of how many times we travelled between nodes.  After finding the target we then return the shortest distance by following the pathway travelled only once.  This effectively eliminates all dead ends.
However, this approach was not sufficient.  Although it met the initial requirements and explanation of the problem, it did not satisfy the testing requirements which included labyrinths with multiple possible solutions and circular paths. There is also a requirement to make it back to the original coordinates within a shorter amount of time than is sometimes possible with the above algorithm.

Therefore for the next attempt I will use the left hand rule only when needed, otherwise I will use a brute force attempt to find the shortest distance to the target when possible.  If after finding the control room the shortest known distance is still not sufficient, then kirk will continue travelling unknown territory using the left hand rule until a quicker route is found.