I completely changed the approach of finding the shortest way back to using a limited brute force attempt, without knowing the entire graph.
However, this still was not efficient enough to handle the challenge.

After further research, I will attempt to use the A* algorithm =)
It is amazing how much more difficult things are made by not knowing the entire graph, and also having larger cavities within the labyrinth, as well as circular and multiple solutions.  For all these reasons, I believe the A* algorithm will prove to be the best.